refactor CPU testing facilities
